move UserGrades page
