+++ frontend/src/pages/Participant/UserGrades/UserGrades.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { Provider } from 'react-redux';
-import { store } from '../../app/store';
+import { store } from '../../../app/store';
 import UserGrades from './UserGrades';
 
 describe('UserGrades', () => {
